Hit Detection added.
Created a collision component that deals with hit detection. Collision Component is added to the Base Weapon Class.
Created an Anim Notify State called, "ANS_CollisionState" that is used by the collision component to decide when to enable and disable collision detection.
Created an event dispatcher called, OnHit. It is dispatched by the collision component for every Pawn hit, it sends the hit result on every dispatch.
Base Weapon is subscribed to OnHit and deals point damage to the Pawn hit.
Combat Character is subscribed to receiving point damage event.