Multiple Attack Types.
Created a new function in Base Weapon called , Get Action Montages. It takes an action type enum as input and returns the corresponding montages array.
Refactored perform attack and perform action function to utilize the Get Action Montages function.
For the charge attack, created a new custom event in character that tracks for how long the input key is held down. If held down for the required time set by the ChargeAttackTime variable, then the charge attack fires.
For sprint attack, checked the player's movement speed mode before attacking. If it is sprinting then it launches the sprint attack.

BUG: Character keeps on sprinting even after the sprint attack. It should revert to jogging speed mode.