Stamina UI Element
Created 2 new widget blueprints. One for the main HUD and one for the progress bar of a stat. Also created a new player controller to add the UI to.
The stat bar widget only contains a progress bar. The stat bar subscribes to the "OnCurrentStatValueUpdated" Event from the stats component and it updates the progress bar whenever the event is fired by the stat component. The stat component fires this event every time a stat value is set.
The main HUD widget has a canvas panel. The stat bar widget is attached to this panel. The stat type is set to stamina for this stat bar.