Stats cost and health UI.
Decreased the various stats' values on certain actions and implemented logic so that charge attacks and dodges may only occur if the player has enough stamina.
Added a health Stat Bar to the player viewport as well.
Added checking for death on stat value updated event for health instead of just checking when the player receives point damage.