Implemented Targeting System
Created a new actor component called TargetingComponent. It handles the functionality the functionality to lock on to a pawn by casting a sphere in the direction the player's camera is facing. If the cast returns a pawn, then the TargetingComponent checks whether the hit pawn can locked on to or not through an interface function. If the interface is Target Interface is implemented by the pawn then it may be locked on. When the character is locked on to a pawn, then it's movement is oriented towards the camera.
Created a new animation blendspace for the tough sword, with directions.
Created a new widget class called WB_LockOn, it only contains a white dot image. This widget was added to the test dummy. It's visibility is set to false by default, enabled when locked on through the Target Interface and disabled when un-targeted.