Fixed engine crashing.
ABP_Manny was causing the engine to crash. The engine would no longer open due to this file so it had to be deleted.
Deleted ABP_Manny and created a new animation blueprint, ABP_CombatCharacter from scratch. Created a new BlendSpace for the Katana and added it to the new animation blueprint. Switched from Manny to Quin.